move template to .cpp
