+++ src/memory/gpdbwrappers.h
@@ -17,87 +17,6 @@ extern "C" {
 #include <string>
 
 namespace ya_gpdb {
-namespace detail {
-
-template <bool Throws, typename Func, typename... Args>
-auto wrap(Func &&func, Args &&...args) noexcept(!Throws)
-    -> decltype(func(std::forward<Args>(args)...)) {
-
-  using RetType = decltype(func(std::forward<Args>(args)...));
-
-  // Empty struct for void return type.
-  struct VoidResult {};
-  using ResultHolder = std::conditional_t<std::is_void_v<RetType>, VoidResult,
-                                          std::optional<RetType>>;
-
-  bool success;
-  ErrorData *edata;
-  ResultHolder result_holder;
-
-  PG_TRY();
-  {
-    if constexpr (!std::is_void_v<RetType>) {
-      result_holder.emplace(func(std::forward<Args>(args)...));
-    } else {
-      func(std::forward<Args>(args)...);
-    }
-    edata = NULL;
-    success = true;
-  }
-  PG_CATCH();
-  {
-    MemoryContext oldctx = MemoryContextSwitchTo(TopMemoryContext);
-    edata = CopyErrorData();
-    MemoryContextSwitchTo(oldctx);
-    FlushErrorState();
-    success = false;
-  }
-  PG_END_TRY();
-
-  if (!success) {
-    std::string err;
-    if (edata && edata->message) {
-      err = std::string(edata->message);
-    } else {
-      err = "Unknown error occurred";
-    }
-
-    if (edata) {
-      FreeErrorData(edata);
-    }
-
-    if constexpr (Throws) {
-      throw std::runtime_error(err);
-    }
-
-    if constexpr (!std::is_void_v<RetType>) {
-      return RetType{};
-    } else {
-      return;
-    }
-  }
-
-  if constexpr (!std::is_void_v<RetType>) {
-    return *std::move(result_holder);
-  } else {
-    return;
-  }
-}
-
-template <typename Func, typename... Args>
-auto wrap_throw(Func &&func, Args &&...args)
-    -> decltype(func(std::forward<Args>(args)...)) {
-  return detail::wrap<true>(std::forward<Func>(func),
-                            std::forward<Args>(args)...);
-}
-
-template <typename Func, typename... Args>
-auto wrap_noexcept(Func &&func, Args &&...args) noexcept
-    -> decltype(func(std::forward<Args>(args)...)) {
-  return detail::wrap<false>(std::forward<Func>(func),
-                             std::forward<Args>(args)...);
-}
-} // namespace detail
 
 // Functions that call palloc().
 // Make sure correct memory context is set.
